fix: changed nomenclature from "gate" to "credential"
