+++ apps/website/components/organisms/gates/view/modals/gate-completed.tsx
@@ -98,8 +98,8 @@ export default function GateCompletedModal({ gate, open, handleClose }) {
               }}
             >
               You have completed the{' '}
-              <span style={{ color: '#D083FF' }}>{gate.title}</span> Gate from{' '}
-              <span style={{ color: '#D083FF' }}>{gate.dao.name}</span>.
+              <span style={{ color: '#D083FF' }}>{gate.title}</span> Credential
+              from <span style={{ color: '#D083FF' }}>{gate.dao.name}</span>.
             </Typography>
           </Box>
           <Box
