+++ apps/website/components/organisms/gates/view/tasks/content/meeting_code.tsx
@@ -34,7 +34,7 @@ const MeetingCodeContent = ({
           Submit
         </LoadingButton>
       )}
-      {completed && (
+      {completed && !!updatedAt && (
         <Typography color="#c5ffe3" variant="subtitle2">
           Task completed at {formattedDate}
         </Typography>
