+++ apps/website/components/organisms/gates/view/tasks/content/quiz.tsx
@@ -110,7 +110,7 @@ const QuizContent = ({
           Submit
         </LoadingButton>
       )}
-      {completed && (
+      {completed && updatedAt && (
         <Typography color="#c5ffe3" variant="subtitle2">
           Task completed at {formattedDate}
         </Typography>
