+++ apps/website/components/organisms/gates/view/tasks/content/snapshot.tsx
@@ -29,7 +29,7 @@ const SnapshotContent = ({
           Check Snapshot
         </LoadingButton>
       )}
-      {completed && (
+      {completed && updatedAt && (
         <Typography color="#c5ffe3" variant="subtitle2">
           Task completed at {formattedDate}
         </Typography>
