+++ apps/website/constants/routes.js
@@ -5,14 +5,15 @@ const ROUTES = {
   NEW_USER: '/new-user',
   CREDENTIALS: '/credentials',
   CREDENTIALS_NEW: '/credentials/new',
+  EARNED: '/earned/[id]',
   DAO_PROFILE: '/dao/[id]',
   MY_PROFILE: '/profile',
   PROFILE: '/profile/[username]',
   PROFILE_EDIT: '/profile/edit',
   DAO_EDIT: '/dao/[id]/edit',
   DAO_NEW: '/dao/new',
-  GATE_PROFILE: '/gate/[id]',
-  GATE_NEW: '/gate/new',
+  GATE_PROFILE: '/credential/[id]',
+  GATE_NEW: '/credential/new',
 };
 
 module.exports = {
