+++ apps/website/pages/earned/new.tsx
@@ -1,0 +1,8 @@
+// Commented for future usage
+// import { NewCredentialTemplate } from '../../components/templates/new-credential';
+
+export default function NewCredential() {
+  // return <NewCredentialTemplate />;
+}
+
+NewCredential.auth = true;
